水一项
321

+++ src/main/java/com/example/cinema/blImpl/management/hall/HallServiceImpl.java
@@ -69,22 +69,23 @@ public class HallServiceImpl implements HallService, HallServiceForBl {
     @Override
     public ResponseVO updateHall(HallVO hallVO){
         try {
-            if( !preCheck(hallVO).getSuccess()){
+            if( preCheck(hallVO).getSuccess()){
                 return ResponseVO.buildFailure("该影厅无法修改！");
-            }
-            int hallId = hallVO.getId();
-            Hall hall = hallMapper.selectHallById(hallId);
-            if(hallVO.getColumn() != hall.getColumn() || hallVO.getRow() != hall.getRow()) {
-                ResponseVO responseVO = preCheck(hallVO);
-                if (!responseVO.getSuccess()) {
-                    return responseVO;
+            }else {
+                int hallId = hallVO.getId();
+                Hall hall = hallMapper.selectHallById(hallId);
+                if (hallVO.getColumn() != hall.getColumn() || hallVO.getRow() != hall.getRow()) {
+                    ResponseVO responseVO = preCheck(hallVO);
+                    if (!responseVO.getSuccess()) {
+                        return responseVO;
+                    } else {
+                        hallMapper.updateHallById(hallVO);
+                        return ResponseVO.buildSuccess();
+                    }
                 } else {
                     hallMapper.updateHallById(hallVO);
                     return ResponseVO.buildSuccess();
                 }
-            }else{
-                hallMapper.updateHallById(hallVO);
-                return ResponseVO.buildSuccess();
             }
         } catch (Exception e) {
             e.printStackTrace();
@@ -95,8 +96,13 @@ public class HallServiceImpl implements HallService, HallServiceForBl {
     @Override
     public ResponseVO deleteHall(int hallId) {
         try {
-            hallMapper.deleteHallById(hallId);
-            return ResponseVO.buildSuccess();
+            Hall hallVO =getHallById(hallId);
+            if( preCheck1(hallVO).getSuccess()){
+                return ResponseVO.buildFailure("该影厅无法修改！");
+            }else {
+                hallMapper.deleteHallById(hallId);
+                return ResponseVO.buildSuccess();
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return ResponseVO.buildFailure("失败");
@@ -136,6 +142,24 @@ public class HallServiceImpl implements HallService, HallServiceForBl {
         }
     }
 
+    private ResponseVO preCheck1(Hall hallVO){
+        try {
+            int hallId = hallVO.getId();
+            int view = scheduleMapper.selectView();
+            SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd");
+            Date today = simpleDateFormat.parse(simpleDateFormat.format(new Date()));
+            Date endDate = getNumDayAfterDate(today, view);
+            List<ScheduleItem> scheduleItems = scheduleMapper.selectSchedule(hallId,today,endDate);
+            if(scheduleItems.size() != 0){
+                return ResponseVO.buildSuccess();
+            }else {
+                return ResponseVO.buildFailure("此影厅无法修改！");
+            }
+        }catch (Exception e){
+            e.printStackTrace();
+            return ResponseVO.buildFailure("失败");
+        }
+    }
     /**
      * 获得num天后的日期
      */
